+++ apps/animated-landing/src/components/cards/CompetitionCard.tsx
