+++ apps/animated-landing/src/components/cards/LockedTaskCard.tsx
