+++ apps/animated-landing/src/components/cards/SubmissionCard.tsx
